HW5/requirements.txt
tensorflow
numpy
matplotlib
keras
h5py
sklearn
Pillow